# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Eltuvchi API",
        "description": "Eltuvchi platformasi uchun API hujjatlari",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/super-admin/dispatchers": {
            "get": {
                "tags": [
                    "Admin Dispatcher"
                ],
                "summary": "Get list of all dispatchers",
                "operationId": "401758a2543078161f5870bdee8b903d",
                "responses": {
                    "200": {
                        "description": "List of dispatchers",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/AdminDispatcher"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Admin Dispatcher"
                ],
                "summary": "Create new dispatcher",
                "operationId": "b95d3de4178835032b52aff6e3a63d13",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "taxopark_id",
                                    "full_name"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "taxopark_id": {
                                        "type": "integer"
                                    },
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "details": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dispatcher created"
                    }
                }
            }
        },
        "/api/super-admin/dispatchers/{dispatcher}": {
            "get": {
                "tags": [
                    "Admin Dispatcher"
                ],
                "summary": "Get single dispatcher",
                "operationId": "756a548aaccc420b3d59f78d086dab0c",
                "parameters": [
                    {
                        "name": "dispatcher",
                        "in": "path",
                        "description": "Dispatcher ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dispatcher details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AdminDispatcher"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Admin Dispatcher"
                ],
                "summary": "Update dispatcher",
                "operationId": "69318ed80334831069cee5bbd92ca492",
                "parameters": [
                    {
                        "name": "dispatcher",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "full_name": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "boolean"
                                    },
                                    "details": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dispatcher updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Admin Dispatcher"
                ],
                "summary": "Deactivate dispatcher (soft delete)",
                "operationId": "614576d67cc43e41e7a3d24ea54ab5d1",
                "parameters": [
                    {
                        "name": "dispatcher",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dispatcher deactivated"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminDispatcher": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "user_id": {
                        "type": "integer"
                    },
                    "taxopark_id": {
                        "type": "integer"
                    },
                    "full_name": {
                        "type": "string"
                    },
                    "status": {
                        "type": "boolean"
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "user": {
                        "type": "object"
                    },
                    "taxopark": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum authentication using Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}